+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Upcoming (v0.5.11)
 
 ### Features
+* Added ThorTiffImagingExtractor for reading TIFF files produced via Thor [#395](https://github.com/catalystneuro/roiextractors/pull/395)
 
 ### Fixes
 * Use tifffile.imwrite instead of tifffile.imsave for TiffImagingExtractor: [#390](https://github.com/catalystneuro/roiextractors/pull/390)
@@ -11,7 +12,8 @@
 
 ### Improvements
 * Use `pyproject.toml` for project metadata and installation requirements [#382](https://github.com/catalystneuro/roiextractors/pull/382)
-
+* Added `__repr__` and  methods to ImagingExtractor for better display in terminals and Jupyter notebooks [#393](https://github.com/catalystneuro/roiextractors/pull/393) and [#396](https://github.com/catalystneuro/roiextractors/pull/396)
+* Removed deprecated np.product from the library [#397](https://github.com/catalystneuro/roiextractors/pull/397)
 
 # v0.5.10 (November 6th, 2024)